Get JDBC to sink data to Postgres
Modified the Python Producer script to produce Avro serialized messages
so that the JDBC connector could work. However the producer service
will exit with status 1 for the first time (as it is not checking wether
the broker is up), you can still solve this by getting the service up
again with `docker-compose up -d producer`. After this, data should start
to appear in the topic quickstart-events and then sink to the Postgres
database.